Create PyDates
The accompanying Python code uses the "datetime" module to demonstrate a number of date-related actions.

The 'datetime.datetime.now()' function is used to first get the current date and time. By using the'strftime()' method to display the current date in several forms, it illustrates the adaptability of date formatting.


The 'timedelta' function from the 'datetime' module is then used to add a certain number of days to the current date. This makes manipulating and calculating dates simple.

The code also shows how to use the strptime() method to transform date strings into "datetime" objects. This is helpful for processing dates that have been stored as strings or when working with user input.

The program also determines the number of days that separate two dates by deducting one from the other. For calculating times between events or durations, this functionality is useful.

The code demonstrates accessing the year, month, and day separately using the appropriate attributes of a datetime object in order to extract particular parts of a date.

Last but not least, the application asks the user to enter the start and end dates and then figures out how many days there are in between. The total claim amount is then calculated using a per-diem rate.

This Python software provides a comprehensive tutorial for working with dates, covering date formatting, math, handling user input, and computing durations for various uses.